backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
pydantic==1.10.7
python-multipart==0.0.6
Pillow==10.2.0
numpy==2.2.4
